pyproject.toml
[project]
name = "browserbase-haystack"
version = "0.0.4"
authors = [
  { name="Browserbase", email="info@browserbase.com" },
]
description = "Browserbase Haystack Fetcher"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
  "haystack-ai",
  "browserbase"
]

[project.urls]
Homepage = "https://browserbase.com"
Issues = "https://github.com/browserbase/haystack/issues"
Source = "https://github.com/browserbase/haystack"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
